Fix compilation issues with MSVC
Well-written code should be compiler-independent; it produces working
software regardless of the compiler used build it. Different compilers
usually produce different erros and warnings; the use of more than one
compiler to build the software should therefore lead to a wider coverage
of errors and warnings and thus also to better code quality.

+++ include/BlockCodec.hpp
@@ -105,7 +105,7 @@ namespace image
     inline uint8_t BlockCodec::findNearest(const ColorTable & color, HighColor ref)
     {
         const auto nearest([ref](auto x, auto y){ return distance(x, ref) < distance(y, ref); });
-        return std::distance(color.begin(), std::min_element(color.begin(), color.end(), nearest));
+        return static_cast<uint8_t>(std::distance(color.begin(), std::min_element(color.begin(), color.end(), nearest)));
     }
 
     template<typename OutputIterator>
